Crate moid

‏
جدولة الزمنية لتنفيذ الدوال المتكرر في خيوط مستقلة قابلة للإيقاف

‏

‏نماذج لاستخدام المكتبة:

‏

use moid;

let handle = moid::repeat(0.25, || println!("أُنَفَذُ كُلُ رُبْع ثَانيِة")).unwrap();
//‎ أوامر أخرى ...
handle.cancel();

‏

‏في المثال السابق ينفذ البرنامج دالة كل ربع ثانية في خيط مستقل، ويستمر الخيط في تنفيذ الدالة إلى أن ينتهي البرنامج الرئيسي أو يستوقف بتنفيذ دالة cancel على قبضة الخيط. فيما يلي مثال لتنفيذ دالة ٦ مرات بعد حلول ثانيتين على إنشاء الخيط. سيحاول الخيط تنفيذ الدالة في ظرف ١٥٠ ملي ثانية ويستغرق ما تبقى من زمن في إنتظار الظرف التالي. تنفيذ join على قبضة الخيط توقف الخيط المنفذ إلى أن ينتهي التكرار (كما هو الحال مع JoinHandle).

‏

use moid;

let mut counter = 0;
let handle = moid::MoidSpec::new()
    .sleep(0.15)
    .delay(2)
    .inclusive()
    .count(6)
    .repeat(move || {
        println!("{}", counter);
        counter += 1;
    })
    .unwrap();
//‎ أوامر أخرى ...
handle.join();

‏

أوجه القصور: لا يمكن استيقاف خيط التكرار أثناء تنفيذ دالته؛ وذلك بسبب عدم وجود طريقة لإجبار الخيوط على التوقف في std::thread. قصور آخر هو عدم قدرة المكتبة على جدولة مدة الإنتظار لأكثر من شهر .

‏

Copyright

Copyright © 2015 Josh Morin JoshMorin@gmx.com

This program is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.

You should have received a copy of the GNU General Public License along with this program. If not, see http://www.gnu.org/licenses/.

Structs

MoidHandle
‎‫قبضة خيط التكرار

MoidSpec
‎وصف لخيط التكرار

Functions

delay
‎‫نفذ الدالة work في خيط منفصل بعد delay ثانية.

nrepeat
‎‫كرر الدالة work ‏count مرات على الأكثر في خيط منفصل كل sleep ثانية.

repeat
‎‫كرر الدالة work في خيط منفصل كل sleep ثانية.

spin
‎‫كرر تنفيذ الدالة work في خيط منفصل.